refactor: move aws secret functionality to it's own file
