+++ src/database-handling/get-country.ts
@@ -1,8 +1,5 @@
 const { createClient } = require("@supabase/supabase-js");
-const {
-  SecretsManagerClient,
-  GetSecretValueCommand,
-} = require("@aws-sdk/client-secrets-manager");
+const getSecret = require("../aws/aws-secret");
 
 // THIS NOW WORKS???
 // const dotenv = require("dotenv");
@@ -18,28 +15,6 @@ const client = new SecretsManagerClient({
   region: "eu-west-2",
 });
 
-// Define a function to get the secret
-async function getSecret() {
-  let response;
-
-  try {
-    response = await client.send(
-      new GetSecretValueCommand({
-        SecretId: secret_name,
-        VersionStage: "AWSCURRENT", // VersionStage defaults to AWSCURRENT if unspecified
-      })
-    );
-
-    // Parse the SecretString to get the actual secret
-    const secret = JSON.parse(response.SecretString);
-    return secret; // Return the secret
-  } catch (error) {
-    console.error("Error retrieving secret:", error);
-    throw error; // Re-throw to handle it later
-  }
-}
-
-// Use the secret in an async IIFE
 // Use the secret in an async IIFE
 const getCountries = async () => {
   try {
